Integrate the afxgroup collaborator PRs (#1-3) into main alongside the existing
software_icd fixes:
- SPIR-V endianness auto-detect from the magic word (accept BE or LE streams).
- ogles2_vk: do not fall back to ogles2vk_LookupProcAddr() (returns APICALL
  trampolines, not raw PFN_vkXxx) - return NULL for unknown names instead.
- ogles2_vk: proc-addr ABI, varying linkage, VkFormat 37, VMA-limits fixes.

Conflicts resolved keeping BOTH sides: the software_vk shader-module ref-count
(UAF) fix + the Vulkan Y-down viewport convention, AND the ogles2_vk vsync/
glFinish (FIFO present pacing) work, AND the afxgroup changes. Merged software
ICD builds clean; ogles2vk_exec.c syntax-checks (ogles2vk_main.c needs the
Warp3DNova GPU SDK header, absent from the CI image).

+++ loader/src/loader_dispatch.c
@@ -70,8 +70,9 @@ VkResult APICALL Loader_vkCreateInstance(struct VulkanIFace *Self,
         VkResult result = pfn(pCreateInfo, pAllocator, pInstance);
         if (result != VK_SUCCESS)
         {
-            IExec->DebugPrintF("[vulkan.library] ICD %s: vkCreateInstance failed (%d)\n",
-                               icd->libraryPath, (int)result);
+            /* RawDoFmt-style: %s for strings, %ld for longs (%d is unsupported). */
+            IExec->DebugPrintF("[vulkan.library] ICD %s: vkCreateInstance failed (%ld)\n",
+                               icd->libraryPath, (long)result);
             continue;
         }
